requirements.txt
python-dotenv
openai
openpyxl
pillow
pandas
pdf2image
python-docx
markdown
beautifulsoup4
